// client/src/app/interfaces/data.ts
export interface ClinVarVarinat {
	uid: number;
	title: string;
	start?: number;
	stop?: number;
	significance?: {
		reviewStatus?: string;
		description?: string;
		lastEvaluated?: string;
	};
	condition?: string;
}